+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [0.3.1] - 2023-03-02
+
+### Added
+- Default handlers that can be attached to every Action created from `kernel.send`
+- Comm Manager that is attached to every Action in `kernel.send`
+  
 ## [0.3.0] - 2023-02-24
 
 ### Changed